chore: remove redundant code and fix component strucuture
