+++ src/components/Dashboard/UrlListItem.tsx
@@ -1,5 +1,7 @@
 import Link from 'next/link';
 import { LiaStopwatchSolid } from 'react-icons/lia';
+import { MdOutlineContentCopy } from 'react-icons/md';
+import { TbTrash } from 'react-icons/tb';
 import { useMutation } from 'react-query';
 
 import { TINY_SITE } from '@/constants/url';
@@ -9,12 +11,37 @@ import { deleteUrlApi } from '@/services/api';
 import { UrlType } from '@/types/url.types';
 import formatDate from '@/utils/formatDate';
 
-import { CopyButton, DeleteButton } from './Buttons';
+import Button from '../Button';
+
+const DeleteButton: React.FC<{ isLoading: boolean; onDelete: () => void }> = ({ isLoading, onDelete }) => (
+    <Button
+        loading={isLoading}
+        onClick={onDelete}
+        className={`flex items-center justify-center w-8 h-8 rounded transition active:scale-95 ${
+            isLoading ? 'text-gray-600 bg-gray-100' : 'text-red-500 hover:text-red-600 hover:bg-red-100'
+        }`}
+        testId="delete-button"
+    >
+        {!isLoading && <TbTrash className="w-5 h-5" />}
+    </Button>
+);
+
+const CopyButton: React.FC<{ onCopy: () => void }> = ({ onCopy }) => (
+    <Button
+        className="rounded-full p-1.5 transition-all duration-75 hover:scale-105 hover:bg-blue-100 active:scale-95"
+        onClick={onCopy}
+        testId="copy-button"
+    >
+        <span className="sr-only">Copy</span>
+        <MdOutlineContentCopy className="text-black" />
+    </Button>
+);
 
 interface UrlListItemProps {
     url: UrlType;
     copyButtonHandler: (url: string) => void;
 }
+
 const UrlListItem: React.FC<UrlListItemProps> = ({ url, copyButtonHandler }) => {
     const { userData } = useAuthenticated();
     const deleteMutation = useMutation({
@@ -30,14 +57,14 @@ const UrlListItem: React.FC<UrlListItemProps> = ({ url, copyButtonHandler }) => 
     const handleDelete = () => deleteMutation.mutate({ id: url.id, userId: Number(userData?.data?.id) });
 
     return (
-        <li className="rounded-xl bg-white px-2 py-2 sm:pl-4 sm:py-4 flex flex-col overflow-hidden">
+        <div className="rounded-xl bg-white px-2 py-2 sm:pl-4 sm:py-4 flex flex-col overflow-hidden">
             <div className="flex items-center gap-2 w-full">
                 <div className="flex flex-col w-full max-w-[calc(100%-4rem)] sm:max-w-[calc(100%-5rem)]">
                     <Link
                         href={`${TINY_SITE}/${url.shortUrl}`}
                         target="_blank"
                         rel="noreferrer"
-                        className="truncate text-sm font-semibold text-black sm:text-base underline-offset-2  hover:underline"
+                        className="truncate text-sm font-semibold text-black sm:text-base underline-offset-2 hover:underline"
                     >
                         {TINY_SITE}/{url.shortUrl}
                     </Link>
@@ -71,7 +98,8 @@ const UrlListItem: React.FC<UrlListItemProps> = ({ url, copyButtonHandler }) => 
                     <DeleteButton isLoading={deleteMutation.isLoading} onDelete={handleDelete} />
                 </div>
             </div>
-        </li>
+        </div>
     );
 };
+
 export default UrlListItem;
