+++ src/components/Navbar/index.tsx
@@ -25,15 +25,21 @@ const Navbar: React.FC = () => {
 
     const toggleMobileMenu = () => setIsMobileMenuOpen((prev) => !prev);
 
-    const profileButton = !isLoggedIn ? (
-        <button onClick={handleProfileClick} className="text-white focus:outline-none sm:hidden z-50">
-            <FaRegUser size={24} />
-        </button>
-    ) : (
-        <button onClick={toggleMobileMenu} className="text-white focus:outline-none sm:hidden z-50">
-            {isMobileMenuOpen ? <FaTimes size={24} /> : <FaBars size={24} />}
-        </button>
-    );
+    const ProfileButton: React.FC = () => {
+        if (!isLoggedIn) {
+            return (
+                <button onClick={handleProfileClick} className="text-white focus:outline-none sm:hidden z-50">
+                    <FaRegUser size={24} />
+                </button>
+            );
+        }
+
+        return (
+            <button onClick={toggleMobileMenu} className="text-white focus:outline-none sm:hidden z-50">
+                {isMobileMenuOpen ? <FaTimes size={24} /> : <FaBars size={24} />}
+            </button>
+        );
+    };
 
     return (
         <nav className="p-4 h-[8vh] mt-3.5 flex items-center justify-between">
@@ -42,7 +48,7 @@ const Navbar: React.FC = () => {
                 <span className="text-white text-2xl font-bold">RDS</span>
             </Link>
             <div className="flex items-center">
-                {profileButton}
+                <ProfileButton />
                 <DesktopMenu
                     isLoading={isLoading}
                     isLoggedIn={isLoggedIn}
